+++ core-ng/src/main/java/core/framework/internal/kafka/ProducerMetrics.java
@@ -12,7 +12,8 @@ import java.util.Map;
  */
 public class ProducerMetrics implements Metrics {
     private final String name;
-    Metric requestSizeAvg; // The average size of requests sent
+    Metric requestSizeAvg; // The average size of all requests in the window for a broker.
+    Metric requestSizeMax; // The maximum size of any request sent in the window for a broker
     private Metric requestRate; // The number of batch requests sent per second, one batch request contains multiple messages
     private Metric outgoingByteRate; // The number of outgoing bytes sent to all servers per second
 
@@ -27,9 +28,14 @@ public class ProducerMetrics implements Metrics {
             Double value = (Double) requestSizeAvg.metricValue();
             if (!value.isNaN()) stats.put(statName("request_size_avg"), value);     // avg value can be NaN, refer to https://cwiki.apache.org/confluence/pages/viewpage.action?pageId=95652345
         }
+        if (requestSizeMax != null) {
+            Double value = (Double) requestSizeMax.metricValue();
+            if (!value.isNaN()) stats.put(statName("request_size_max"), value);
+        }
         if (outgoingByteRate != null) stats.put(statName("outgoing_byte_rate"), (Double) outgoingByteRate.metricValue());
     }
 
+    // refer to https://docs.confluent.io/platform/current/kafka/monitoring.html#per-broker-metrics
     public void set(Map<MetricName, ? extends Metric> kafkaMetrics) {
         for (Map.Entry<MetricName, ? extends Metric> entry : kafkaMetrics.entrySet()) {
             MetricName metricName = entry.getKey();
@@ -37,6 +43,7 @@ public class ProducerMetrics implements Metrics {
                 String name = metricName.name();
                 if ("request-rate".equals(name)) requestRate = entry.getValue();
                 else if ("request-size-avg".equals(name)) requestSizeAvg = entry.getValue();
+                else if ("request-size-max".equals(name)) requestSizeMax = entry.getValue();
                 else if ("outgoing-byte-rate".equals(name)) outgoingByteRate = entry.getValue();
             }
         }